requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
black==25.1.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.2
daphne==4.1.2
distlib==0.3.9
Django==5.1.7
django-celery-beat==2.7.0
django-channels==0.7.0
django-ckeditor==6.7.2
django-environ==0.12.0
django-filter==25.1
django-image-uploader-widget==1.0.1
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-modeltranslation==0.19.12
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
factory_boy==3.3.3
Faker==37.1.0
filelock==3.18.0
flake8==7.1.2
hyperlink==21.0.0
identify==2.6.9
idna==3.10
importlib_metadata==8.6.1
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
kombu==5.5.0
mccabe==0.7.0
msgpack==1.1.0
mypy==1.15.0
mypy-extensions==1.0.0
nodeenv==1.9.1
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
pre_commit==4.2.0
prompt_toolkit==3.0.50
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
pyOpenSSL==25.0.0
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-django==4.11.1
pytest-factoryboy==2.7.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==78.1.0
six==1.17.0
sorl-thumbnail==12.11.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
virtualenv==20.29.3
wcwidth==0.2.13
zipp==3.21.0
zope.interface==7.2
